+++ include/stream/StreamReader.h
@@ -28,7 +28,7 @@ namespace BlizzardDatabaseLib {
           virtual pos_type seekoff(off_type off, std::ios_base::seekdir dir, std::ios_base::openmode which = std::ios_base::in) override
           {
             if(dir == std::ios_base::cur)
-              gbump(off);
+              gbump(static_cast<int>(off));
             else if(dir == std::ios_base::end)
               setg(_buf, _buf + _size + off, _buf + _size);
             else if(dir == std::ios_base::beg)
